AMBR-632 extract service
